requirements.txt
mysql-connector-python==8.0.18
Flask==1.1.1
Flask-Bcrypt==0.7.1
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SocketIO==2.9.3
